Clean up Go module dependencies
The go.sum files included hashes for modules which seem unrelated to the
project. Because golangci-lint was installed using "GO11MODULE=on go
get", it was added as a dependency, which is not necessary. This means
that third-party projects which want to use this one as a module now end
up with all these hashes in their own go.sum file, which is not
desirable.

This commit simply runs "go mod tidy" with the desired version of go
(1.12 for this project) to clean up dependencies. We also change the
installation method for golangci-lint so that the binary is downloaded
directly.

We also update the Azure CI pipeline definition to verify that future
PRs include up-to-date go.mod and go.sum files. The CI pipeline
definition is simplified by leveraging improved support for recent Go
versions.